Move flatten_label from pyk.kast to pyk.kastManip
